fixed exit code
